+++ packages/mobx-state-tree/__tests__/core/reference.test.ts
@@ -15,7 +15,8 @@ import {
     Instance,
     SnapshotOrInstance,
     isAlive,
-    destroy
+    destroy,
+    castToReferenceSnapshot
 } from "../../src"
 
 test("it should support prefixed paths in maps", () => {
@@ -45,7 +46,7 @@ test("it should support prefixed paths in maps", () => {
     expect(getSnapshot(store)).toEqual({
         user: "18",
         users: { "17": { id: "17", name: "Michel" }, "18": { id: "18", name: "Noa" } }
-    }) // TODO: better typings
+    } as SnapshotOut<typeof store>)
 })
 test("it should support prefixed paths in arrays", () => {
     const User = types.model({
@@ -122,10 +123,10 @@ test("it should resolve refs during creation, when using path", () => {
     })
     unprotect(s)
     reaction(() => s.entries.reduce((a, e) => a + e.price, 0), v => values.push(v))
-    s.entries.push(cast({ book: s.books[0] }))
+    s.entries.push({ book: castToReferenceSnapshot(s.books[0]) })
     expect(s.entries[0].price).toBe(4)
     expect(s.entries.reduce((a, e) => a + e.price, 0)).toBe(4)
-    const entry = BookEntry.create({ book: s.books[0] }) // N.B. ref is initially not resolvable!
+    const entry = BookEntry.create({ book: castToReferenceSnapshot(s.books[0]) }) // N.B. ref is initially not resolvable!
     s.entries.push(entry)
     expect(s.entries[1].price).toBe(4)
     expect(s.entries.reduce((a, e) => a + e.price, 0)).toBe(8)
@@ -153,7 +154,7 @@ test("it should resolve refs over late types", () => {
         books: [{ id: "3", price: 2 }]
     })
     unprotect(s)
-    s.entries.push(cast({ book: s.books[0] }))
+    s.entries.push({ book: castToReferenceSnapshot(s.books[0]) })
     expect(s.entries[0].price).toBe(4)
     expect(s.entries.reduce((a, e) => a + e.price, 0)).toBe(4)
 })
@@ -181,10 +182,10 @@ test("it should resolve refs during creation, when using generic reference", () 
     })
     unprotect(s)
     reaction(() => s.entries.reduce((a, e) => a + e.price, 0), v => values.push(v))
-    s.entries.push(cast({ book: s.books[0] }))
+    s.entries.push({ book: castToReferenceSnapshot(s.books[0]) })
     expect(s.entries[0].price).toBe(4)
     expect(s.entries.reduce((a, e) => a + e.price, 0)).toBe(4)
-    const entry = BookEntry.create({ book: s.books[0] }) // can refer to book, even when not part of tree yet
+    const entry = BookEntry.create({ book: castToReferenceSnapshot(s.books[0]) }) // can refer to book, even when not part of tree yet
     expect(getSnapshot(entry)).toEqual({ book: "3" })
     s.entries.push(entry)
     expect(values).toEqual([4, 8])
@@ -247,7 +248,7 @@ test("self reference with a late type", () => {
             books: types.array(Book)
         })
         .actions(self => {
-            function addBook(book: typeof Book.Type | typeof Book.CreationType) {
+            function addBook(book: SnapshotOrInstance<typeof Book>) {
                 self.books.push(book)
             }
             return {
@@ -260,7 +261,7 @@ test("self reference with a late type", () => {
     const book2 = Book.create({
         id: "2",
         genre: "romance",
-        reference: s.books[0]
+        reference: castToReferenceSnapshot(s.books[0])
     })
     s.addBook(book2)
     expect((s.books[1].reference as Instance<typeof Book>).genre).toBe("thriller")
@@ -309,7 +310,7 @@ test("it should fail when reference snapshot is ambiguous", () => {
     )
     unprotect(store)
     // first update the reference, than create a new matching item! Ref becomes ambigous now...
-    store.selected = cast(1) // valid assignment
+    store.selected = 1 as any // valid assignment
     expect(store.selected).toBe(store.boxes[0]) // unambigous identifier
     let err!: Error
     autorun(() => store.selected, {
@@ -318,7 +319,7 @@ test("it should fail when reference snapshot is ambiguous", () => {
         }
     })
     expect(store.selected).toBe(store.boxes[0]) // unambigous identifier
-    store.arrows.push(cast({ id: 1, name: "oops" }))
+    store.arrows.push({ id: 1, name: "oops" })
     expect(err.message).toBe(
         "[mobx-state-tree] Cannot resolve a reference to type '(Box | Arrow)' with id: '1' unambigously, there are multiple candidates: /boxes/0, /arrows/1"
     )
@@ -477,9 +478,9 @@ test("References are non-nullable by default", () => {
         expect(() => store.ref).toThrow(
             "[mobx-state-tree] Failed to resolve reference '4' to type 'AnonymousModel' (from node: /ref)"
         )
-        store.maybeRef = cast(3) // valid assignment
+        store.maybeRef = 3 as any // valid assignment
         expect(store.maybeRef).toBe(store.todo)
-        store.maybeRef = cast(4) // valid assignment
+        store.maybeRef = 4 as any // valid assignment
         expect(() => store.maybeRef).toThrow(
             "[mobx-state-tree] Failed to resolve reference '4' to type 'AnonymousModel' (from node: /maybeRef)"
         )
@@ -517,7 +518,9 @@ test("References in recursive structures", () => {
             function addFolder(data: SnapshotOrInstance<typeof Folder>) {
                 const folder3 = Folder.create(data)
                 getRoot<typeof Storage>(self).putFolderHelper(folder3)
-                self.children.push(Tree.create({ data: folder3, children: [] }))
+                self.children.push(
+                    Tree.create({ data: castToReferenceSnapshot(folder3), children: [] })
+                )
             }
             return { addFolder }
         })
@@ -750,11 +753,11 @@ test("should serialize references correctly", () => {
     })
     expect(Array.from(s.mies.keys())).toEqual(["7", "8"])
 
-    s.ref = cast(8)
+    s.ref = 8 as any
     expect(s.ref!.id).toBe(8) // resolved from number
     expect(getSnapshot(s).ref).toBe(8) // ref serialized as number
 
-    s.ref = cast("7") // resolved from string
+    s.ref = "7" as any // resolved from string
     expect(s.ref!.id).toBe(7) // resolved from string
     expect(getSnapshot(s).ref).toBe("7") // ref serialized as string (number would be ok as well)
 
@@ -762,7 +765,7 @@ test("should serialize references correctly", () => {
     expect(s.ref.id).toBe(8) // resolved from instance
     expect(getSnapshot(s).ref).toBe(8) // ref serialized as number
 
-    s.ref = cast("9") // unresolvable
+    s.ref = "9" as any // unresolvable
     expect(getSnapshot(s).ref).toBe("9") // snapshot preserved as it was unresolvable
 
     s.mies.set(9 as any, {
@@ -802,7 +805,7 @@ test("#1052 - Reference returns destroyed model after subtree replacing", () => 
             },
             select(todo: Instance<typeof Todo>) {
                 self.last = todo
-                self.lastWithId = cast(todo.id)
+                self.lastWithId = todo.id as any
             }
         }))