+++ packages/mobx-state-tree/__tests__/core/union.test.ts
@@ -5,7 +5,8 @@ import {
     getSnapshot,
     applySnapshot,
     setLivelynessChecking,
-    getType
+    getType,
+    SnapshotIn
 } from "../../src"
 
 const createTestFactories = () => {